requirements.txt
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

google-generativeai>=0.8.5
groq>=0.4.0

chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.30.0
langchain>=0.1.0

pypdf>=3.15.0
pydub>=0.25.1
pytesseract>=0.3.10
Pillow>=10.0.0
fpdf>=1.7.2

pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.13

numpy>=1.24.0
tqdm>=4.66.0
langdetect>=1.0.9
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0

streamlit-mic-recorder>=0.0.7

pandas>=2.0.0
